+++ src/features/todos/index.tsx
@@ -1,0 +1,9 @@
+import TaskManagementBoard from './components/task-management-board/TaskManagementBoard';
+
+export default function Todos() {
+    return (
+        <>
+            <TaskManagementBoard />
+        </>
+    );
+}
