+++ src/styles/base/index.scss
@@ -1,1 +1,1 @@
-@import '../../components/buttons/primary-button/primary-button.scss';
+@import '../../components/buttons/primary-button/primary-button';
