--- src/styles/components/index.scss
@@ -1,1 +1,0 @@
-@import '../../features/task-management-board/components/task-management-board.scss';
